chore: move application-options to internal/types
