+++ generators/base-application/application-options-all.d.ts
