+++ generators/base-application/internal/types/application-options.d.ts
